Change build from Visual Studio compiler to MingW-32

+++ KinectProject/CoreLib/types/Point3f.cpp
@@ -9,7 +9,7 @@ Point3f::Point3f()
     m_pos[0] = m_pos[1] = m_pos[2] = 0.0f;
 }
 
-Point3f::Point3f(double x, double y, double z)
+Point3f::Point3f(float x, float y, float z)
 {
     this->set(x, y, z);
 }
@@ -19,7 +19,7 @@ Point3f::Point3f(const Point3f& other)
     *this = other;
 }
 
-void Point3f::set(double x, double y, double z)
+void Point3f::set(float x, float y, float z)
 {
     m_pos[0] = x;
     m_pos[1] = y;
@@ -42,22 +42,22 @@ bool Point3f::operator!=(const Point3f& other) const
     return !operator==(other);
 }
 
-double Point3f::x() const
+float Point3f::x() const
 {
     return m_pos[0];
 }
 
-double Point3f::y() const
+float Point3f::y() const
 {
     return m_pos[1];
 }
 
-double Point3f::z() const
+float Point3f::z() const
 {
     return m_pos[2];
 }
 
-const double* Point3f::dataPtr() const
+const float* Point3f::dataPtr() const
 {
     return m_pos;
 }
@@ -66,9 +66,9 @@ const double* Point3f::dataPtr() const
 // Static methods
 //
 
-double Point3f::distance(const Point3f& point1, const Point3f& point2, DistanceType option)
+float Point3f::distance(const Point3f& point1, const Point3f& point2, DistanceType option)
 {
-    double result = std::numeric_limits<double>::max();
+    float result = std::numeric_limits<float>::max();
 
     if (option == DISTANCE_EUCLIDEAN) {
         result = euclideanDistance(point1, point2);
@@ -77,11 +77,11 @@ double Point3f::distance(const Point3f& point1, const Point3f& point2, DistanceT
     return result;
 }
 
-double Point3f::euclideanDistance(const Point3f& point1, const Point3f& point2)
+float Point3f::euclideanDistance(const Point3f& point1, const Point3f& point2)
 {
-    double dx = point2.x() - point1.x();
-    double dy = point2.y() - point1.y();
-    double dz = point2.z() - point1.z();
+    float dx = point2.x() - point1.x();
+    float dy = point2.y() - point1.y();
+    float dz = point2.z() - point1.z();
     return sqrt(pow(dx, 2) + pow(dy, 2) + pow(dz, 2));
 }
