+++ KinectProject/CoreLib/types/Point3f.h
@@ -11,25 +11,25 @@ public:
         DISTANCE_EUCLIDEAN
     };
 
-    static double distance(const Point3f& point1, const Point3f& point2, DistanceType option = DISTANCE_EUCLIDEAN);
+    static float distance(const Point3f& point1, const Point3f& point2, DistanceType option = DISTANCE_EUCLIDEAN);
 
     Point3f();
-    explicit Point3f(double x, double y, double z);
+    explicit Point3f(float x, float y, float z);
     Point3f(const Point3f& other);
-    const double* dataPtr() const;
+    const float *dataPtr() const;
     Point3f& operator=(const Point3f& other);
     bool operator==(const Point3f& other) const;
     bool operator!=(const Point3f& other) const;
 
-    double x() const;
-    double y() const;
-    double z() const;
+    float x() const;
+    float y() const;
+    float z() const;
 
 private:    
-    static double euclideanDistance(const Point3f& point1, const Point3f& point2);
-    void set(double x, double y, double z);
+    static float euclideanDistance(const Point3f& point1, const Point3f& point2);
+    void set(float x, float y, float z);
 
-    double m_pos[3]; // x, y, z
+    float m_pos[3]; // x, y, z
 };
 
 } // End Namepsace
